requirements.txt
# -- Packages that should be installed through anaconda. --
# torch~=1.13.1
# torchvision~=0.14.1
# -- Packages the should be installed through pip. --
click~=8.1.7
numpy~=1.26.4
Pillow~=10.2.0
setuptools~=69.2.0
pytest~=8.1.1
packaging~=24.0
einops~=0.7.0
scipy~=1.12.0
filetype~=1.2.0
scikit-learn~=1.4.1.post1
pytorch-lightning~=2.2.1
torchmetrics~=1.3.1
tqdm~=4.66.2
PyYAML~=6.0.1
albumentations~=1.4.1
tensorboard~=2.16.2
warmup_scheduler_pytorch~=0.1.2
lmdb~=1.4.1
matplotlib~=3.8.3
git+https://github.com/facebookresearch/segment-anything.git
timm~=0.9.16
jpegio~=0.2.8
yacs~=0.1.8
seaborn~=0.13.0
opencv-python-headless~=4.9.0.80
networkx~=3.4.2